service-chatbot-main/requirements.txt
fastapi
pydantic
psycopg2
google.generativeai
pandas
Pillow
python-multipart
uvicorn
streamlit
cachetools
pydantic_settings
PrettyTable
openpyxl